move search inputs creation out of graphqlbackend
This moves the creation of the SearchInputs type out of graphqlbackend
and into `internal/search/run`.
